+++ cmd/branchProtection_i_test.go
@@ -105,125 +105,6 @@ func (suite *BranchProtectionSuite) cleanup() {
 	}
 }
 
-func (suite *BranchProtectionSuite) TestGetChecksForWorkflowContentWithListSyntax() {
-	verifier := BranchProtectionVerifier{}
-	definition, err := verifier.parseWorkflowDefinition(`
-name: CI Build
-on:
-  - push
-jobs:
-  build:
-    runs-on: ubuntu-latest
-`)
-	suite.NoError(err)
-	suite.Contains(definition.JobsNames, "build")
-	suite.Contains(definition.Trigger, "push")
-	suite.Contains(definition.Name, "CI Build")
-
-}
-
-func (suite *BranchProtectionSuite) TestGetChecksForWorkflowContentWithMapSyntax() {
-	verifier := BranchProtectionVerifier{}
-	definition, err := verifier.parseWorkflowDefinition(`
-name: CI Build
-on:
-  push:
-jobs:
-  build:
-    runs-on: ubuntu-latest
-`)
-	suite.NoError(err)
-	suite.Contains(definition.JobsNames, "build")
-	suite.Contains(definition.Trigger, "push")
-	suite.Contains(definition.Name, "CI Build")
-
-}
-
-func (suite *BranchProtectionSuite) TestGetChecksForWorkflowContentWithJobName() {
-	verifier := BranchProtectionVerifier{}
-	definition, err := verifier.parseWorkflowDefinition(`
-name: CI Build
-on:
-  push:
-jobs:
-  build:
-    name: My-Job
-    runs-on: ubuntu-latest
-`)
-	suite.NoError(err)
-	suite.Contains(definition.JobsNames, "My-Job")
-
-}
-
-func (suite *BranchProtectionSuite) TestGetChecksForWorkflowContentWithMatrixBuild() {
-	verifier := BranchProtectionVerifier{}
-	definition, err := verifier.parseWorkflowDefinition(`
-name: CI Build
-on:
-  push:
-jobs:
-  build:
-    strategy:
-      matrix:
-        a: [ "1", "2"]
-        b: [ "3", "4" ]
-    name: Build with A ${{ matrix.a }} and B ${{ matrix.b }}
-    runs-on: ubuntu-latest
-`)
-	suite.NoError(err)
-	suite.Contains(definition.JobsNames, "Build with A 1 and B 3")
-	suite.Contains(definition.JobsNames, "Build with A 1 and B 4")
-	suite.Contains(definition.JobsNames, "Build with A 2 and B 3")
-	suite.Contains(definition.JobsNames, "Build with A 2 and B 4")
-}
-
-func (suite *BranchProtectionSuite) TestGetChecksForWorkflowContentWithMatrixBuildWithMultiplesParameters() {
-	verifier := BranchProtectionVerifier{}
-	definition, err := verifier.parseWorkflowDefinition(`
-name: CI Build
-on:
-  push:
-jobs:
-  build:
-    strategy:
-      matrix:
-        a:
-         - id: 1
-           num: 10
-         - id: 2
-           num: 20
-        b: [ "3" ]
-    name: Build with id ${{ matrix.id }}, num ${{matrix.num}} and B ${{ matrix.b }}
-    runs-on: ubuntu-latest
-`)
-	suite.NoError(err)
-	suite.Contains(definition.JobsNames, "Build with id 1, num 10 and B 3")
-	suite.Contains(definition.JobsNames, "Build with id 2, num 20 and B 3")
-}
-
-func (suite *BranchProtectionSuite) TestGetChecksForWorkflowContentWithMatrixBuildWithMultiplesParametersAnNoName() {
-	verifier := BranchProtectionVerifier{}
-	definition, err := verifier.parseWorkflowDefinition(`
-name: CI Build
-on:
-  push:
-jobs:
-  build:
-    strategy:
-      matrix:
-        a:
-         - id: 1
-           num: 10
-         - id: 2
-           num: 20
-        b: [ "3" ]
-    runs-on: ubuntu-latest
-`)
-	suite.NoError(err)
-	suite.Contains(definition.JobsNames, "build (1, 10, 3)")
-	suite.Contains(definition.JobsNames, "build (2, 20, 3)")
-}
-
 type TestHasWorkflowPushOrPrTriggerCase struct {
 	trigger        []string
 	expectedResult bool
